# apps/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TinyClaw API",
    "version": "1",
    "description": "HTTP API for the TinyClaw personal AI assistant."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:4310",
      "description": "Local dev server"
    }
  ],
...
  "components": {
    "schemas": {
...
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "WorkerLogsResponse": {
        "type": "object",
        "properties": {
          "worker": {
            "type": "string"
          },
          "lines": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "worker",
          "lines"
        ],
        "additionalProperties": {}
      },
...
      "UserContextStatusResponse": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "UpdateUserContextRequest": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
      "InitUserContextResponse": {
        "type": "object",
        "properties": {},
        "additionalProperties": {}
      },
...
    },
...
  },
  "paths": {
...
    "/v1/workers/{name}/{action}": {
      "post": {
        "tags": [
          "Workers"
        ],
        "summary": "Control a worker",
        "operationId": "workerAction",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "name",
            "in": "path"
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "start",
                "stop",
                "restart"
              ]
            },
            "required": true,
            "name": "action",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Worker action succeeded",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workers/{name}/logs": {
      "get": {
        "tags": [
          "Workers"
        ],
        "summary": "Get worker logs",
        "operationId": "getWorkerLogs",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "name",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "lines",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Worker logs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkerLogsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workers/{name}/clear-logs": {
      "post": {
        "tags": [
          "Workers"
        ],
        "summary": "Clear worker logs",
        "operationId": "clearWorkerLogs",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "name",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Worker logs cleared",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "ok"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/user/context": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Get USER.md status",
        "operationId": "getUserContext",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "enum": [
                "true",
                "false"
              ]
            },
            "required": false,
            "name": "content",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "User context status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserContextStatusResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Write USER.md",
        "operationId": "writeUserContext",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserContextRequest"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "User context saved"
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/user/context/init": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Initialize USER.md template",
        "operationId": "initUserContext",
        "responses": {
          "201": {
            "description": "User context initialized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitUserContextResponse"
                }
              }
            }
          },
          "500": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}